+++ src/boyboy/core/cpu/cpu.cpp
@@ -10,6 +10,7 @@
 #include <cstdint>
 
 #include "boyboy/common/log/logging.h"
+#include "boyboy/common/utils.h"
 #include "boyboy/core/cpu/instructions.h"
 #include "boyboy/core/cpu/instructions_table.h"
 #include "boyboy/core/profiling/profiler_utils.h"
@@ -120,17 +121,26 @@ void Cpu::set_register(Reg16Name reg, uint16_t value)
     }
 }
 
+void Cpu::set_halted(bool halted)
+{
+    if (halted != halted_) {
+        log::debug("CPU HALT mode {}", halted ? "entered" : "exited");
+    }
+    halted_ = halted;
+}
+
 uint8_t Cpu::step()
 {
     BB_PROFILE_SCOPE(profiling::FrameTimer::Cpu);
 
-    interrupt_handler_.service();
+    uint8_t cycles = interrupt_handler_.service();
+    cycles_ += cycles;
 
     if (halted_) {
         // TODO: This should be 0 cycles, but until we have a clock, we use 4 cycles
         // to keep other components ticking
         cycles_ += 4;
-        return 4;
+        return cycles + 4;
     }
 
 #ifdef DISASSEMBLY_LOG
@@ -145,7 +155,7 @@ uint8_t Cpu::step()
         instr_type = InstructionType::CBPrefixed;
     }
 
-    uint8_t cycles = execute(opcode, instr_type);
+    cycles += execute(opcode, instr_type);
 
     // IME is enabled after the instruction following EI
     if (ime_scheduled_ &&
@@ -162,6 +172,16 @@ uint8_t Cpu::fetch()
     BB_PROFILE_START(profiling::HotSection::CpuFetch);
     uint8_t result = read_byte(registers_.pc++);
     BB_PROFILE_STOP(profiling::HotSection::CpuFetch);
+
+    if (halt_bug_) {
+        // when halt bug occurs we don't increment PC
+        registers_.pc--;
+        halt_bug_ = false;
+        log::debug(
+            "CPU HALT bug handled, not advancing PC={}", utils::PrettyHex{get_pc()}.to_string()
+        );
+    }
+
     return result;
 }