+++ src/boyboy/core/io/timer.cpp
@@ -2,10 +2,24 @@
  * @file timer.cpp
  * @brief Timer operations for BoyBoy emulator.
  *
+ * On how the timer circuit works:
+ * Basically, TIMA is incremented with the falling edge of a logic AND between TAC enable and
+ * certain bits of the internal system counter based on TAC selected frequency.
+ * This means two things:
+ *     1) Any change in DIV, TAC.enable or TAC.freq that triggers the falling edge detector,
+ *        will increase TIMA. i.e. TAC.enable && DIV(TAC.freq) 1 -> 0; TIMA++
+ *     2) If resetting DIV before the tested bit is set, TIMA will never be incremented no matter
+ *        how many ticks.
+ *
+ * For more information check Pan Docs and/or The Cycle-Accurate Game Boy Docs.
+ *
  * @license GPLv3 (see LICENSE file)
  */
 
 #include "boyboy/core/io/timer.h"
+
+#include <algorithm>
+#include <cstdint>
 
 #include "boyboy/core/io/registers.h"
 
@@ -13,32 +27,31 @@ namespace boyboy::core::io {
 
 void Timer::tick(uint16_t cycles)
 {
-    // Update DIV register (if not stopped)
-    if (!stopped_) {
-        div_counter_ += cycles;
-        if (div_counter_ >= Frequency::DivIncrement) {
-            div_ += div_counter_ / Frequency::DivIncrement;
-            div_counter_ %= Frequency::DivIncrement;
-        }
+    if (stopped_) {
+        return;
     }
 
-    // Check if timer is enabled
-    if ((tac_ & Flags::TimerEnable) != 0) {
-        // Determine the frequency
-        uint16_t freq = get_frequency();
+    // Increment at a rate of 4 T-cycles (1 M-cycle)
+    while (cycles > 0) {
 
-        // Update TIMA register
-        tima_counter_ += cycles;
-        while (tima_counter_ >= freq) {
-            tima_counter_ -= freq;
-            if (tima_ == 0xFF) {
-                tima_ = tma_; // Reset TIMA to TMA on overflow
-                request_interrupt_(cpu::Interrupts::Timer);
-            }
-            else {
-                tima_++;
-            }
+        auto cur_cycles = std::min(cycles, uint16_t{4});
+
+        if (interrupt_scheduler_.scheduled) {
+            interrupt_scheduler_.update(cur_cycles);
         }
+        if (tima_reload_scheduler_.scheduled) {
+            tima_reload_scheduler_.update(cur_cycles);
+        }
+
+        // Update DIV counter
+        increment_div_counter(cur_cycles);
+
+        // Check if TIMA overflow happened and schedule
+        if (tima_overflow_) {
+            schedule_overflow();
+        }
+
+        cycles -= cur_cycles;
     }
 }
 
@@ -46,7 +59,7 @@ uint8_t Timer::read(uint16_t addr) const
 {
     switch (addr) {
         case IoReg::Timer::DIV:
-            return div_;
+            return div_counter_ >> 8;
         case IoReg::Timer::TIMA:
             return tima_;
         case IoReg::Timer::TMA:
@@ -62,16 +75,17 @@ void Timer::write(uint16_t addr, uint8_t value)
 {
     switch (addr) {
         case IoReg::Timer::DIV:
-            div_ = 0; // writing any value resets DIV
+            // writing any value resets the counter
+            set_div_counter(0);
             break;
         case IoReg::Timer::TIMA:
-            tima_ = value;
+            set_tima(value);
             break;
         case IoReg::Timer::TMA:
-            tma_ = value;
+            set_tma(value);
             break;
         case IoReg::Timer::TAC:
-            tac_ = value & 0x07; // only lower 3 bits are used
+            set_tac(value);
             break;
         default:
             break;
@@ -85,12 +99,13 @@ void Timer::set_interrupt_cb(cpu::InterruptRequestCallback callback)
 
 void Timer::reset()
 {
-    div_ = 0;
     tima_ = 0;
     tma_ = 0;
     tac_ = 0;
-    div_counter_ = 0;
+    div_counter_ = DivCounterStartValue;
     tima_counter_ = 0;
+    interrupt_scheduler_.reset();
+    tima_reload_scheduler_.reset();
     stopped_ = false;
 }
 
@@ -102,7 +117,6 @@ void Timer::start()
 void Timer::stop()
 {
     stopped_ = true;
-    div_ = 0;
     div_counter_ = 0;
 }